Now correctly adding and removing photos

+++ src/control/actions/update_pet.php
@@ -4,7 +4,10 @@ require_once(dirname(__FILE__) . "/action.php");
 require_once(dirname(__FILE__) . "/../api/pet.php");
 require_once(dirname(__FILE__) . "/../api/user.php");
 
-$parameters = initAction(['petId', 'name', 'location', 'description', 'removePhotos']);
+if (isset($_POST['removePhotos']))
+	$parameters = initAction(['petId', 'name', 'location', 'description', 'removePhotos']);
+else
+	$parameters = initAction(['petId', 'name', 'location', 'description']);
 
 $petId = API\updatePet($parameters['petId'], $parameters['name'], $parameters['location'], $parameters['description']); // update pet
 
@@ -19,14 +22,18 @@ for ($i = 0; $i < sizeof($_FILES['photos']['name']); $i++) {
 	$originalPath = "../../images/petPictures/" . $photoId . ".jpg";
 	move_uploaded_file($tmpPath, $originalPath);
 
+	print_r($_FILES);
+	print_r($parameters);
 	if ($_FILES['photos']['name'][$i] == $parameters['profilePhoto']) {
 		copy($originalPath, "../../images/petProfilePictures/" . $petId . ".jpg");
 	}
 }
 
-for ($i = 0; $i < sizeof($parameters['removePhotos']); $i++) {
-	API\removePetPhoto($parameters['removePhotos'][$i]);
-	unlink( $_SERVER['DOCUMENT_ROOT'] . '/images/petPictures/' . $parameters['removePhotos'][$i] . '.jpg');
+if (isset($parameters['removePhotos'])) {
+	for ($i = 0; $i < sizeof($parameters['removePhotos']); $i++) {
+		API\removePetPhoto($parameters['removePhotos'][$i]);
+		unlink($_SERVER['DOCUMENT_ROOT'] . '/images/petPictures/' . $parameters['removePhotos'][$i] . '.jpg');
+	}
 }
 
 Router\sendTo(getRootUrl() . "/pet/" . $parameters['petId']);
